ClassificationMetrics

Documentation for ClassificationMetrics.

ClassificationMetrics.ConfusionMatrix

ClassificationMetrics.PredictionResults

ClassificationMetrics.confusion_matrix

ClassificationMetrics.ConfusionMatrix — Type

struct ConfusionMatrix{T1, T2}

Confusion matrix struct that holds the matrix and label names.

label_set: Class labels

matrix: Confusion matrix

ClassificationMetrics.PredictionResults — Type

struct PredictionResults{T1, T2}

Holds TP, FP, TN, and FN values for further metric calculation.

label_set: Class labels

results: A $4 \times n$ matrix with rows corresponding to TP, FN, FP and TN.

ClassificationMetrics.confusion_matrix — Function

confusion_matrix(predicted, actual[, label_set]; sort_labels=false)
confusion_matrix(predicted::OneHotLike, actual::OneHotLike, label_set)

Compute confusion matrix. Predicted and actual can be provided as vectors of labels or in the one-hot encoded form. Optionally, label set can be provided explicitly.